+++ src/main/java/com/nirima/jenkins/plugins/docker/builder/DockerBuilderPublisher.java
@@ -1,6 +1,7 @@
 package com.nirima.jenkins.plugins.docker.builder;
 
 import com.github.dockerjava.api.DockerClient;
+import com.github.dockerjava.api.command.BuildImageCmd;
 import com.github.dockerjava.api.command.PushImageCmd;
 import com.github.dockerjava.api.exception.DockerException;
 import com.github.dockerjava.api.model.AuthConfigurations;
@@ -59,11 +60,14 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.regex.Pattern;
 
 import static com.nirima.jenkins.plugins.docker.utils.JenkinsUtils.fixEmpty;
+import static com.nirima.jenkins.plugins.docker.utils.JenkinsUtils.splitAndFilterEmptyMap;
 import static com.nirima.jenkins.plugins.docker.utils.JenkinsUtils.splitAndTrimFilterEmptyList;
 import static com.nirima.jenkins.plugins.docker.utils.LogUtils.printResponseItemToListener;
 import static org.apache.commons.lang.StringUtils.isEmpty;
@@ -128,6 +132,9 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
 
     @CheckForNull
     private List<String> tags;
+	
+    @CheckForNull
+    private Map<String, String> buildArgs;
 
     public final boolean pushOnSuccess;
 
@@ -209,6 +216,27 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
     }
 
     @CheckForNull
+    public Map<String, String> getBuildArgs() {
+        return fixEmpty(buildArgs);
+    }
+
+    @DataBoundSetter
+    public void setBuildArgs(Map<String, String> buildArgs) {
+        this.buildArgs = fixEmpty(buildArgs);
+    }
+
+    @Nonnull
+    public String getBuildArgsString() {
+        final Map<String, String> buildArgsOrNull = getBuildArgs();
+        return buildArgsOrNull == null ? "" : Joiner.on("\n").withKeyValueSeparator("=").join(buildArgsOrNull);
+    }
+
+    @DataBoundSetter
+    public void setBuildArgsString(String buildArgsString) {
+        setBuildArgs(splitAndFilterEmptyMap(buildArgsString, "\n"));
+    }
+
+    @CheckForNull
     public DockerRegistryEndpoint getFromRegistry() {
         return fromRegistry;
     }
@@ -281,14 +309,17 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
         private final TaskListener listener;
         private final FilePath fpChild;
         private final List<String> tagsToUse;
+        @Nonnull
+        private final Map<String, String> buildArgsToUse;
         private final DockerAPI dockerApi;
         private final hudson.model.Run<?, ?> run;
 
-        private Run(hudson.model.Run<?, ?> run, final TaskListener listener, FilePath fpChild, List<String> tagsToUse, DockerAPI dockerApi) {
+        private Run(hudson.model.Run<?, ?> run, final TaskListener listener, FilePath fpChild, List<String> tagsToUse, @Nonnull Map<String, String> buildArgsToUse, DockerAPI dockerApi) {
             this.run = run;
             this.listener = listener;
             this.fpChild = fpChild;
             this.tagsToUse = tagsToUse;
+            this.buildArgsToUse = buildArgsToUse;
             this.dockerApi = dockerApi;
         }
 
@@ -365,12 +396,12 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
             };
             final String imageId;
             try(final DockerClient client = getClientWithNoTimeout()) {
-                imageId = client.buildImageCmd(tar)
+                BuildImageCmd cmd = client.buildImageCmd(tar)
                         .withNoCache(noCache)
                         .withPull(pull)
-                        .withBuildAuthConfigs(auths)
-                        .exec(resultCallback)
-                        .awaitImageId();
+                        .withBuildAuthConfigs(auths);
+                buildArgsToUse.forEach(cmd::withBuildArg);
+                imageId = cmd.exec(resultCallback).awaitImageId();
                 if (imageId == null) {
                     throw new AbortException("Built image id is null. Some error occured");
                 }
@@ -444,6 +475,7 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
     @Override
     public void perform(hudson.model.Run<?, ?> run, FilePath workspace, Launcher launcher, TaskListener listener) throws InterruptedException, IOException {
         final List<String> expandedTags = expandTags(run, workspace, listener);
+		final Map<String, String> buildArgsToUse = Optional.ofNullable(getBuildArgs()).orElseGet(Collections::emptyMap);
         String expandedDockerFileDirectory = dockerFileDirectory;
         try {
             expandedDockerFileDirectory = TokenMacro.expandAll(run, workspace, listener, this.dockerFileDirectory);
@@ -451,7 +483,7 @@ public class DockerBuilderPublisher extends Builder implements SimpleBuildStep {
             listener.getLogger().println("Couldn't macro expand docker file directory " + dockerFileDirectory);
             e.printStackTrace(listener.getLogger());
         }
-        new Run(run, listener, new FilePath(workspace, expandedDockerFileDirectory), expandedTags, getDockerAPI(launcher)).run();
+        new Run(run, listener, new FilePath(workspace, expandedDockerFileDirectory), expandedTags, buildArgsToUse, getDockerAPI(launcher)).run();
     }
 
     @Override